+++ src/notes.ts
@@ -12,6 +12,111 @@ export interface SortAttribute {
 }
 
 /**
+ * Returns attributes at the given path for the given note.
+ *
+ * A path can either be a single attribute name or multiple names separated by
+ * a period (".").
+ *
+ * If the path is a single attribute name, then this will return all attributes
+ * defined on the given note having that name.
+ *
+ * If the path consists of multiple names, then the last name refers to an
+ * attribute and all other names in the path refer to relations. This will
+ * fetch all notes targeted by those relations, then return all attributes from
+ * those targeted notes having the attribute name.
+ *
+ * For example, the path "x.y.z" would mean:
+ *
+ * 1. Fetch all notes targeted by relation "x" of the given note.
+ * 2. Fetch all notes targeted by relation "y" of all notes from step 1.
+ * 3. Return all "z" attributes of all notes from step 2.
+ *
+ * A name can either be the name of an attribute or one of the special names
+ * listed below. These special names refer to a note's properties instead of
+ * user-defined attributes. Properties are returned in the form of a label.
+ *
+ * - $id or $noteId: The note's ID.
+ * - $type: The note's type (as listed under Note Info such as "text").
+ * - $mime: The note's content type (as listed under Note Info such as
+ *   "text/html").
+ * - $title: The note's title.
+ * - $contentSize: The size of the note's content in bytes.
+ * - $dateCreated: The note's creation date in UTC and RFC 3339 format.
+ * - $dateModified: The note's modification date in UTC and RFC 3339 format.
+ *
+ * If an attribute is not found, then an empty array is returned.
+ *
+ * If a relation refers to a note that does not exist, then it is ignored.
+ */
+export async function getAttributes(
+	note: NoteShort,
+	path: string
+): Promise<Attribute[]> {
+	if (!path) {
+		return [];
+	}
+
+	const parts = path.split(".");
+	if (parts.length > 1) {
+		const attributes = [];
+		const targetPath = parts.slice(1).join(".");
+		for (const target of await note.getRelationTargets(parts[0])) {
+			const targetAttributes = await getAttributes(target, targetPath);
+			attributes.push(...targetAttributes);
+		}
+
+		return attributes;
+	}
+
+	let value: string | undefined;
+	switch (path) {
+		case "$id":
+		case "$noteId":
+			value = note.noteId;
+			break;
+		case "$type":
+			value = note.type;
+			break;
+		case "$mime":
+			value = note.mime;
+			break;
+		case "$title":
+			value = note.title;
+			break;
+		case "$contentSize":
+			value = `${(await note.getNoteComplement()).contentLength}`;
+			break;
+		case "$dateCreated":
+			value = (await note.getNoteComplement()).utcDateCreated;
+			break;
+		case "$dateModified":
+			value = (await note.getNoteComplement()).combinedUtcDateModified;
+			break;
+	}
+	if (value !== undefined) {
+		return [{ type: "label", value }];
+	}
+
+	return note.getAttributes(undefined, path);
+}
+
+/**
+ * Returns the first value of the attribute referenced by the given path that
+ * belong to the given note or notes targeted by the given note's relations.
+ *
+ * See getAttributes for more details.
+ *
+ * If no attributes are found, then an empty string is returned.
+ */
+export async function getAttributeValue(
+	note: NoteShort,
+	name: string
+): Promise<string> {
+	const attributes = await getAttributes(note, name);
+	return attributes[0]?.value || "";
+}
+
+/**
  * Returns the URL for a note's cover image or undefined if it has none.
  */
 export async function getCoverUrl(