+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## 1.2.0 - Unreleased
 
+- Add support for finding attributes of notes targeted by a note's relations, similar to that supported by [Trilium's search engine](https://github.com/zadam/trilium/wiki/Search#advanced-use-cases).
+  - Wherever an attribute name can be specified, a path can now be specified instead.
+  - A path consists of one or more names separated by a period (`.`). The last name in the path must be an attribute name. All other names in the path must be relation names.
+  - For example:
+    - `name` would find attributes named `name` defined on a note.
+    - `employee.name` would find attributes named `name` defined on all notes targeted by the `employee` relation defined on a note.
+    - `company.employee.name` would find attributes named `name` defined on all notes targeted by the `employee` relation defined on all notes targeted by the `company` relation defined on a note.
+- Add support for finding properties of a note.
+  - Wherever an attribute name can be specified, there are now some special names (prefixed with `$`) which refer to a note's properties instead of its user-defined attributes.
+  - `$id` and `$noteId` are the note's ID.
+  - `$type` is the note type (for example, `text`).
+  - `$mime` is the note's content type (for example, `text/html`).
+  - `$title` is the note's title.
+  - `$contentSize` is the size of the note's content in bytes.
+  - `$dateCreated` is the note's creation date and time in UTC and RFC 3339 format (`YYYY-MM-DD hh:mm:ss.sssZ`).
+  - `$dateModified` is the note's modification date and time in UTC and RFC 3339 format.
 - Add a `separator` attribute setting for controlling how multiple values for a single attribute are separated:
   - `separator=newline` inserts a newline between values, resulting in one value per line.
   - `separator=comma` inserts a comma and space between values.